DestroysAnnouncementStates action implemented
