+++ app/services/cancels_course_membership/destroys_announcement_states.rb
@@ -1,20 +1,16 @@
 module Services
   module Actions
-    class DestroysMembership
+    class DestroysAnnouncementStates
       extend LightService::Action
 
       expects :membership
 
       executed do |context|
         membership = context[:membership]
-        membership.destroy
-      end
-    end
 
-    class DestroysFlaggedUsers
-      extend LightService::Action
-
-      executed do |context|
+        AnnouncementState.for_course(membership.course)
+          .for_user(membership.user)
+          .destroy_all
       end
     end
   end
